requirements.txt
python-dotenv==0.21.0
fastapi==0.85.0
pydantic==1.10.2
sqlalchemy==1.4.41
psycopg2-binary==2.9.3
uvicorn[standard]==0.18.3
alembic==1.8.1